Add new flow interaction blocks as well as specification tests.
FlowCrossbar models an n to m connection block which replicates
incoming FlowObject based messages to a target set that is
runtime managed.

FlowSource is a placeholder for a FlowObject origin. For now
creates a random Sentiment object each second and sends it to
a configurable target.

Spec tests check basic properties of all flowNetwork components.
Based on Akka testkit integration taken from
http://blog.xebia.com/2012/10/01/testing-akka-with-specs2/