Node updates interrupt dragging #14
When a node receives a position update while being dragged it is moved back. With elements receiving regular updates this isn't acceptable. Now suspending updates for element while dragging. This is fine as the move will cause a property update for the element anyways so it won't be outdated afterwards.