Work around #3 and #4 for now.
Instead of a proper client specific history replay or handling
of reconnects simply trigger a global configuration detection.
This will cause the whole flow configuration to be pushed
to all connected clients which is not desirable for a real
system but works for now.

Also added error detection on the EventStream which cleans the
drawing surface in case of errors. Probably should do more
like disable all controls or something but it's fine for now.