Fix some minor issues as well as work around a race condition on load.
Unfortunately they akka backend can be faster responding to the config
detection request than the streaming connection to the client can be
opened. This caused messages to be discarded (aka objects not being
synced correctly) on reload. A small delay was added to work around
that. The real solution is to fix the way we provide data for each
client. Another issue was ordering. As nodes and connections are in
no way ordered a connection might arrive before its nodes. To prevent
it from being dropped on the floor the client now retries remote
connections a few times if the objects they relate to are missing.