Introduce configuration update rate limiting.
Not exactly matching the description but still fixes #11 .
The rate limiting prevents the flow network from flooding
the network with configuration updates killing the
responsiveness of the app by overloading the network and the
client browser.

With this patch configuration updates above 4 per second are
discarded. As this in theory might lead to important configuration
updates being lost updates created in reaction to outside
configuration changes aren't limited.