Introduce FlowTrace sink
Part of #10. Element which displays a table of
n (by default 5) previously received messages in a
table inside the element.

The js has special behavior for this element
dependent on the NodeType. In theory this could
easily be generalized but this would probably not
be worth the work right now.