Make FlowFrequency operational and add some eyecandy.
Don't bother saving the FlowObject. We can't really use it anyways.
If we limit ourselves to the string representation things are
pretty straight-forward. Also fixed broken sort. I should get rid
of the notion of anything being in-place in scala ;) Note that
while the list growing/shrinking is exposed growing is broken
once the list has entries > 1

Display in GUI is in form of a table listing position, word and value.

Also fixed groupings and choose nicer icons in node selector.